src/ml_workflow/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
pillow>=9.0.0
numpy>=1.24.0

# Google Cloud
google-cloud-storage>=2.10.0
fsspec>=2023.1.0
gcsfs>=2023.1.0

# Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0